requirements.txt
aiohttp==3.5.4
fuzzywuzzy==0.18.0
matplotlib==3.4.0
matplotlib==3.7.0
networkx==2.6.3
networkx==3.1
numpy==1.21.5
numpy==1.23.5
pandas==1.3.5
PyDriller==2.5.1
pymannkendall==1.4.2
pyquery==1.4.3
pytest==7.1.2
python_Levenshtein==0.21.1
python_Levenshtein==0.23.0
Requests==2.31.0
scikit_learn==1.0.2
scikit_learn==1.3.0
scipy==1.7.3
scipy==1.10.1
seaborn==0.13.1
setuptools==65.6.3
statsmodels==0.13.2
statsmodels==0.14.0
tigramite==5.0.0.3
venn==0.1.3
~tatsmodels==0.13.5
